requirements.txt
discord.py==2.3.2
python-dotenv>=0.19.0
openai>=1.0.0
requests>=2.26.0
notion-client>=2.0.0 